Refactor support-stripe to be entirely clickable
