+++ addon/templates/components/support-stripe.hbs
@@ -1,21 +1,8 @@
-{{#if link}}
-  <a href={{link}} target="_blank">
-    <div class="support-stripe__wrapper">
-      {{yield (hash
-      link=(component 'nypr-ui/support-stripe/link')
-      text=(component 'nypr-ui/support-stripe/heading')
-      )}}
-    </div>
+<div class="support-stripe__wrapper">
+  {{yield (hash
+    label=(component 'nypr-ui/support-stripe/label')
+    text=(component 'nypr-ui/support-stripe/heading')
+  )}}
+</div>
 
-    <div class="support-stripe__image" style={{image}}></div>
-  </a>
-{{else}}
-  <div class="support-stripe__wrapper">
-    {{yield (hash
-      link=(component 'nypr-ui/support-stripe/link')
-      text=(component 'nypr-ui/support-stripe/heading')
-    )}}
-  </div>
-
-  <div class="support-stripe__image" style={{image}}></div>
-{{/if}}
+<div class="support-stripe__image" style={{image}}></div>
