+++ addon/templates/components/support-stripe/label.hbs
