+++ app/components/nypr-ui/support-stripe/label.js
@@ -1,0 +1,1 @@
+export { default } from 'nypr-ui/components/support-stripe/label';
