--- app/components/nypr-ui/support-stripe/link.js
@@ -1,1 +1,0 @@
-export { default } from 'nypr-ui/components/support-stripe/link';
